add booking methods

+++ src/main/java/org/example/hotelbookingapp/entity/Hotel.java
@@ -35,7 +35,7 @@ public class Hotel {
     @Embedded
     private Address address;
 
-    private byte starsCount;
+    private int starsCount;
 
     @Builder.Default
     @OneToMany(mappedBy = "hotel", cascade = CascadeType.ALL)
